refactor: inline WS helpers into useAgentStateWebSocket; remove agentStateService
