+++ console/src/hooks/useAgentStateWebSocket.ts
@@ -1,6 +1,55 @@
 import { useCallback, useEffect, useRef, useState } from "react";
-import { createAgentStateWebSocketUrl, AGENT_STATE_WEB_SOCKET_RECONNECT_DELAY_MS, parseAgentStateResponse } from "../agentStateService";
-import type { AgentStateResponse } from "../agentStateService";
+
+export const AGENT_STATE_WEB_SOCKET_PATH = "/console/api/ws";
+export const AGENT_STATE_WEB_SOCKET_RECONNECT_DELAY_MS = 2_000;
+
+export const createAgentStateWebSocketUrl = (baseHref: string): string => {
+  return new URL(AGENT_STATE_WEB_SOCKET_PATH, baseHref).toString();
+};
+
+export type AgentStateResponse = {
+  error?: string
+  niconama?: {
+    type?: string
+    meta?: {
+      title?: string
+      url?: string
+      start?: number
+      total?: {
+        listeners?: number
+        gift?: number
+        ad?: number
+        comments?: number
+      }
+    }
+  }
+  canSpeak?: boolean
+  currentGame?: {
+    name?: string
+    state?: Record<string, unknown>
+  } | null
+  nGram?: number
+  nGramRaw?: number
+  speech?: {
+    speech?: string
+    silent?: boolean
+  }
+  speechHistory?: Array<{
+    id?: string
+    speech?: string
+    nGram?: number
+    nGramRaw?: number
+  }>
+};
+const INVALID_AGENT_STATE_RESPONSE_ERROR = "配信状態の応答形式が不正です。";
+
+export const parseAgentStateResponse = (responseText: string): AgentStateResponse => {
+  try {
+    return JSON.parse(responseText) as AgentStateResponse;
+  } catch {
+    throw new SyntaxError(INVALID_AGENT_STATE_RESPONSE_ERROR);
+  }
+};
 
 type UseAgentStateWebSocketParams = {
   onMessage: (response: AgentStateResponse) => void;
@@ -31,7 +80,7 @@ export function useAgentStateWebSocket({ onMessage, onError, enabled = true }: U
       return;
     }
 
-    const socket = new WebSocket(createAgentStateWebSocketUrl(`wss://${window.location.host}`));
+      const socket = new WebSocket(createAgentStateWebSocketUrl(`wss://${window.location.host}`));
     websocketRef.current = socket;
 
     const scheduleReconnect = () => {
